Modulariza utilidades y optimiza ajuste de color
