+++ agents_tareas
@@ -9,7 +9,7 @@
 9. [x] Agregar controles de reproducción: Play/Stop (barra espaciadora), Adelantar, Atrasar e Inicio.
 10. [x] Añadir opciones de aspecto 16:9 y 9:16, con soporte para pantalla completa y supermuestreo.
 11. [x] Implementar menú inferior con dropdowns de Instrumento y Familia siempre visibles, y panel desplegable para personalizar color y figura por familia.
-12. Optimizar rendimiento y modularizar el código con comentarios claros para facilitar el desarrollo incremental.
+12. [x] Optimizar rendimiento y modularizar el código con comentarios claros para facilitar el desarrollo incremental.
 13. [x] Implementar pruebas unitarias básicas para las funciones de parseo de archivos MIDI y MusicXML.
 14. [x] Integrar la estructura de pistas con el menú inferior, poblando dinámicamente el dropdown de instrumentos según los archivos cargados.
 15. [x] Renderizar rectángulos estáticos en el canvas representando notas basadas en eventos (subtarea de la 5).
@@ -27,3 +27,5 @@
 27. [x] Permitir restablecer los colores y figuras personalizados de las familias a los valores predeterminados.
 28. [x] Permitir exportar e importar la configuración de familias e instrumentos en archivos JSON.
 29. [x] Crear pruebas unitarias para la exportación e importación de configuraciones.
+30. Refactorizar la lógica de la interfaz en módulos separados y crear pruebas de integración básicas.
+31. Optimizar el renderizado de notas utilizando técnicas de canvas offscreen.
