+++ index.html
@@ -75,6 +75,7 @@
     </div>
   </div>
 
+  <script src="utils.js"></script>
   <script src="script.js"></script>
 </body>
 </html>
